fix: address CI failures from composer-busy gate (issue #78)
- Split PTY payload types out of protocol.ts into pty-protocol.ts so the
  file stays under the ~500-line cap; re-export from protocol.ts to keep
  existing imports working.
- Update api-cmd-runtime unit test to expect the new vendor option passed
  to deliverHostedPrompt.
- Make the behavior harness fake-claude emit a ❯ prompt glyph so the C-layer
  composer-empty gate sees an empty composer, and adjust pty-api-autostart
  to verify teardown via pty-list (delete --force removes the task record).
